pyproject.toml
[project]
name = "opendatasoft"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.4.1",
    "httpx>=0.28.1",
    "ipykernel>=6.30.1",
    "pathlib>=1.0.1",
    "polars>=1.32.3",
    "pyyaml>=6.0.2",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
]
